:art: Make the description extension more generic
We use a description extension in other places in the IG. This one should serve those
purposes too, given that it has the ID "description". An alternate approach would
be to give this a specific id (like study-family-description) as is currently done
by the other description extensions. However, it seems more consistent for users to
be able to refer back to a single description extension rather than having to handle
multiple structure definitions for approximately the same thing. We can specify
the resource-specific semantics in the client resource (as is currently done below).