:bug: Fix sample identifier and id
The FHIR convention for a unique identifier is to use the id field
from Resource. I made it 1..1 rather than the default 0..1 to express
the desire that there be a unique ID present in the Sample.

The identifier field is for the multiple business identifiers. For
example, BioSampleId, the dbGaP Sample ID, and the submitter's sample
ID. I restored "identifier" to its default 0..* cardinality and added
a description.