:t-rex: Align family module with GA4GH pedigree IG
Derive NcpiFamilyRelationship from GA4GH Pedigree Relationship. Since
it is not directly published, adding the dependency requires modifying
build and setup instructions to manually install the IG to the FHIR
cache.

This also updates the descriptions and instructions to match the new
requirements on the relationship code system - it must now be the KIN
ontology because of the required binding on the GA4GH Pedigree
Relationship's codes.